+++ MSAL/src/public/MSALError.h
@@ -126,6 +126,11 @@ extern NSString *MSALHomeAccountIdKey;
 extern NSString *MSALErrorDomain;
 
 /**
+Flag indicate the error returned from client-side throttling.
+ */
+extern NSString *MSALThrottlingCacheHitKey;
+
+/**
  MSALError enum contains all errors that should be considered for handling in runtime.
  */
 typedef NS_ENUM(NSInteger, MSALError)